+++ hapi-fhir-jpaserver-base/src/test/java/ca/uhn/fhir/jpa/provider/dstu3/ResourceProviderDstu3ConceptMapTest.java
@@ -9,15 +9,15 @@ import org.hl7.fhir.dstu3.model.BooleanType;
 import org.hl7.fhir.dstu3.model.CodeType;
 import org.hl7.fhir.dstu3.model.Coding;
 import org.hl7.fhir.dstu3.model.ConceptMap;
+import org.hl7.fhir.dstu3.model.ConceptMap.ConceptMapGroupComponent;
+import org.hl7.fhir.dstu3.model.ConceptMap.SourceElementComponent;
+import org.hl7.fhir.dstu3.model.ConceptMap.TargetElementComponent;
+import org.hl7.fhir.dstu3.model.Enumerations.ConceptMapEquivalence;
 import org.hl7.fhir.dstu3.model.Parameters;
 import org.hl7.fhir.dstu3.model.Parameters.ParametersParameterComponent;
 import org.hl7.fhir.dstu3.model.StringType;
 import org.hl7.fhir.dstu3.model.UriType;
 import org.hl7.fhir.instance.model.api.IIdType;
-import org.hl7.fhir.dstu3.model.ConceptMap.ConceptMapGroupComponent;
-import org.hl7.fhir.dstu3.model.ConceptMap.SourceElementComponent;
-import org.hl7.fhir.dstu3.model.ConceptMap.TargetElementComponent;
-import org.hl7.fhir.dstu3.model.Enumerations.ConceptMapEquivalence;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
@@ -25,7 +25,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.transaction.annotation.Transactional;
 
 import ca.uhn.fhir.rest.api.MethodOutcome;
-import ca.uhn.fhir.rest.client.api.IGenericClient;
 
 public class ResourceProviderDstu3ConceptMapTest extends BaseResourceProviderDstu3Test {
 	private static final Logger ourLog = LoggerFactory.getLogger(ResourceProviderDstu3ConceptMapTest.class);
@@ -129,46 +128,14 @@ public class ResourceProviderDstu3ConceptMapTest extends BaseResourceProviderDst
 	}
 	
 	@Test
-	public void testTranslateWithConcaptMapUrlAndVersion() {
+	public void testTranslateWithConceptMapUrlAndVersion() {
 		
-		//- conceptMap1 v1
-		ConceptMap conceptMap1 = new ConceptMap();
-		conceptMap1.setUrl(CM_URL).setVersion("v1").setSource(new UriType(VS_URL)).setTarget(new UriType(VS_URL_2));
-		
-		ConceptMapGroupComponent group1 = conceptMap1.addGroup();
-		group1.setSource(CS_URL).setSourceVersion("Version 1").setTarget(CS_URL_2).setTargetVersion("Version 2");
-
-		SourceElementComponent element1 = group1.addElement();
-		element1.setCode("11111").setDisplay("Source Code 11111");
-
-		TargetElementComponent target1 = element1.addTarget();
-		target1.setCode("12222").setDisplay("Target Code 12222").setEquivalence(ConceptMapEquivalence.EQUAL);
-		
-		IIdType conceptMapId1 = myConceptMapDao.create(conceptMap1, mySrd).getId().toUnqualifiedVersionless();
-		conceptMap1 = myConceptMapDao.read(conceptMapId1);
-		
-		ourLog.info("ConceptMap: 2 \n" + myFhirCtx.newJsonParser().setPrettyPrint(true).encodeResourceToString(conceptMap1));
-
-		//- conceptMap1 v2
-		ConceptMap conceptMap2 = new ConceptMap();
-		conceptMap2.setUrl(CM_URL).setVersion("v2").setSource(new UriType(VS_URL)).setTarget(new UriType(VS_URL_2));
-		
-		ConceptMapGroupComponent group2 = conceptMap2.addGroup();
-		group2.setSource(CS_URL).setSourceVersion("Version 1").setTarget(CS_URL_2).setTargetVersion("Version 2");
-
-		SourceElementComponent element2 = group2.addElement();
-		element2.setCode("11111").setDisplay("Source Code 11111");
-
-		TargetElementComponent target2 = element2.addTarget();
-		target2.setCode("13333").setDisplay("Target Code 13333").setEquivalence(ConceptMapEquivalence.EQUAL);
-		
-		IIdType conceptMapId2 = myConceptMapDao.create(conceptMap2, mySrd).getId().toUnqualifiedVersionless();
-		conceptMap2 = myConceptMapDao.read(conceptMapId2);
-		
-		ourLog.info("ConceptMap: 2 \n" + myFhirCtx.newJsonParser().setPrettyPrint(true).encodeResourceToString(conceptMap2));
+		String url = "http://url";
+		createConceptMap(url, "v1", "12222", "Target Code 12222");
+		createConceptMap(url, "v2", "13333", "Target Code 13333");
 
 		Parameters inParams = new Parameters();
-		inParams.addParameter().setName("url").setValue(new UriType(CM_URL));
+		inParams.addParameter().setName("url").setValue(new UriType(url));
 		inParams.addParameter().setName("conceptMapVersion").setValue(new StringType("v2"));
 		inParams.addParameter().setName("system").setValue(new UriType(CS_URL));
 		inParams.addParameter().setName("targetsystem").setValue(new UriType(CS_URL_2));
@@ -176,7 +143,6 @@ public class ResourceProviderDstu3ConceptMapTest extends BaseResourceProviderDst
 
 		ourLog.info("Request Parameters:\n" + myFhirCtx.newJsonParser().setPrettyPrint(true).encodeResourceToString(inParams));
 
-		
 		Parameters respParams = ourClient
 			.operation()
 			.onType(ConceptMap.class)
@@ -205,11 +171,287 @@ public class ResourceProviderDstu3ConceptMapTest extends BaseResourceProviderDst
 		assertEquals(CS_URL_2, coding.getSystem());
 		assertEquals("Version 2", coding.getVersion());
 		part = getPartByName(param, "source");
-		assertEquals(CM_URL, ((UriType) part.getValue()).getValueAsString());
+		assertEquals(url, ((UriType) part.getValue()).getValueAsString());
 	}
 	
 	@Test
-	public void testTranslateWithReverseConcaptMapUrlAndVersion() {
+	public void testTranslateWithVersionedConcaptMapUrl_v2() {
+
+		String url = "http://url";
+		createConceptMap(url, "v1", "12222", "Target Code 12222");
+		createConceptMap(url, "v2", "13333", "Target Code 13333");
+		
+		// Call translate with ConceptMap v2.
+		Parameters inParams = new Parameters();
+		inParams.addParameter().setName("url").setValue(new UriType(url));
+		inParams.addParameter().setName("conceptMapVersion").setValue(new StringType("v2"));
+		inParams.addParameter().setName("system").setValue(new UriType(CS_URL));
+		inParams.addParameter().setName("targetsystem").setValue(new UriType(CS_URL_2));
+		inParams.addParameter().setName("code").setValue(new CodeType("11111"));
+
+		ourLog.info("Request Parameters:\n" + myFhirCtx.newJsonParser().setPrettyPrint(true).encodeResourceToString(inParams));
+
+		Parameters respParams = ourClient
+			.operation()
+			.onType(ConceptMap.class)
+			.named("translate")
+			.withParameters(inParams)
+			.execute();
+
+		ourLog.info("Response Parameters\n" + myFhirCtx.newJsonParser().setPrettyPrint(true).encodeResourceToString(respParams));
+
+		// Should return v2 since v2 specified.
+		ParametersParameterComponent param = getParameterByName(respParams, "result");
+		assertTrue(((BooleanType) param.getValue()).booleanValue());
+
+		param = getParameterByName(respParams, "message");
+		assertEquals("Matches found!", ((StringType) param.getValue()).getValueAsString());
+
+		assertEquals(1, getNumberOfParametersByName(respParams, "match"));
+		param = getParametersByName(respParams, "match").get(0);
+		assertEquals(3, param.getPart().size());
+		ParametersParameterComponent part = getPartByName(param, "equivalence");
+		assertEquals("equal", ((CodeType) part.getValue()).getValueAsString());
+		part = getPartByName(param, "concept");
+		Coding coding = (Coding) part.getValue();
+		assertEquals("13333", coding.getCode());
+		assertEquals("Target Code 13333", coding.getDisplay());
+		assertFalse(coding.getUserSelected());
+		assertEquals(CS_URL_2, coding.getSystem());
+		assertEquals("Version 2", coding.getVersion());
+		part = getPartByName(param, "source");
+		assertEquals(url, ((UriType) part.getValue()).getValueAsString());
+
+	}
+
+	@Test
+	public void testTranslateWithVersionedConcaptMapUrl_v1() {
+
+		String url = "http://url";
+		createConceptMap(url, "v1", "12222", "Target Code 12222");
+		createConceptMap(url, "v2", "13333", "Target Code 13333");
+		
+		// Call translate with ConceptMap v1.
+		Parameters inParams = new Parameters();
+		inParams.addParameter().setName("url").setValue(new UriType(url));
+		inParams.addParameter().setName("conceptMapVersion").setValue(new StringType("v1"));
+		inParams.addParameter().setName("system").setValue(new UriType(CS_URL));
+		inParams.addParameter().setName("targetsystem").setValue(new UriType(CS_URL_2));
+		inParams.addParameter().setName("code").setValue(new CodeType("11111"));
+
+		ourLog.info("Request Parameters:\n" + myFhirCtx.newJsonParser().setPrettyPrint(true).encodeResourceToString(inParams));
+
+		Parameters respParams = ourClient
+			.operation()
+			.onType(ConceptMap.class)
+			.named("translate")
+			.withParameters(inParams)
+			.execute();
+
+		ourLog.info("Response Parameters\n" + myFhirCtx.newJsonParser().setPrettyPrint(true).encodeResourceToString(respParams));
+
+		// Should return v1 since v1 specified.
+		ParametersParameterComponent param = getParameterByName(respParams, "result");
+		assertTrue(((BooleanType) param.getValue()).booleanValue());
+
+		param = getParameterByName(respParams, "message");
+		assertEquals("Matches found!", ((StringType) param.getValue()).getValueAsString());
+
+		assertEquals(1, getNumberOfParametersByName(respParams, "match"));
+		param = getParametersByName(respParams, "match").get(0);
+		assertEquals(3, param.getPart().size());
+		ParametersParameterComponent part = getPartByName(param, "equivalence");
+		assertEquals("equal", ((CodeType) part.getValue()).getValueAsString());
+		part = getPartByName(param, "concept");
+		Coding coding = (Coding) part.getValue();
+		assertEquals("12222", coding.getCode());
+		assertEquals("Target Code 12222", coding.getDisplay());
+		assertFalse(coding.getUserSelected());
+		assertEquals(CS_URL_2, coding.getSystem());
+		assertEquals("Version 2", coding.getVersion());
+		part = getPartByName(param, "source");
+		assertEquals(url, ((UriType) part.getValue()).getValueAsString());
+
+	}
+
+	@Test
+	public void testTranslateWithVersionedConcaptMapUrl_NoVersion() {
+
+		String url = "http://url";
+		createConceptMap(url, "v1", "12222", "Target Code 12222");
+		createConceptMap(url, "v2", "13333", "Target Code 13333");
+		
+		// Call translate with no ConceptMap version.
+		Parameters inParams = new Parameters();
+		inParams.addParameter().setName("url").setValue(new UriType(url));
+		inParams.addParameter().setName("system").setValue(new UriType(CS_URL));
+		inParams.addParameter().setName("targetsystem").setValue(new UriType(CS_URL_2));
+		inParams.addParameter().setName("code").setValue(new CodeType("11111"));
+
+		ourLog.info("Request Parameters:\n" + myFhirCtx.newJsonParser().setPrettyPrint(true).encodeResourceToString(inParams));
+
+		Parameters respParams = ourClient
+			.operation()
+			.onType(ConceptMap.class)
+			.named("translate")
+			.withParameters(inParams)
+			.execute();
+
+		ourLog.info("Response Parameters\n" + myFhirCtx.newJsonParser().setPrettyPrint(true).encodeResourceToString(respParams));
+
+		// Should return v2 since v2 is the most recently updated version.
+		ParametersParameterComponent param = getParameterByName(respParams, "result");
+		assertTrue(((BooleanType) param.getValue()).booleanValue());
+
+		param = getParameterByName(respParams, "message");
+		assertEquals("Matches found!", ((StringType) param.getValue()).getValueAsString());
+
+		assertEquals(1, getNumberOfParametersByName(respParams, "match"));
+		param = getParametersByName(respParams, "match").get(0);
+		assertEquals(3, param.getPart().size());
+		ParametersParameterComponent part = getPartByName(param, "equivalence");
+		assertEquals("equal", ((CodeType) part.getValue()).getValueAsString());
+		part = getPartByName(param, "concept");
+		Coding coding = (Coding) part.getValue();
+		assertEquals("13333", coding.getCode());
+		assertEquals("Target Code 13333", coding.getDisplay());
+		assertFalse(coding.getUserSelected());
+		assertEquals(CS_URL_2, coding.getSystem());
+		assertEquals("Version 2", coding.getVersion());
+		part = getPartByName(param, "source");
+		assertEquals(url, ((UriType) part.getValue()).getValueAsString());
+	}
+
+	@Test
+	public void testTranslateWithVersionedConcaptMapUrl_NoVersion_null_v1() {
+
+		String url = "http://url";
+		createConceptMap(url, null, "12222", "Target Code 12222"); // first version is null
+		createConceptMap(url, "v2", "13333", "Target Code 13333");
+				 		
+		// Call translate with no ConceptMap version.
+		Parameters inParams = new Parameters();
+		inParams.addParameter().setName("url").setValue(new UriType(url));
+		inParams.addParameter().setName("system").setValue(new UriType(CS_URL));
+		inParams.addParameter().setName("targetsystem").setValue(new UriType(CS_URL_2));
+		inParams.addParameter().setName("code").setValue(new CodeType("11111"));
+
+		ourLog.info("Request Parameters:\n" + myFhirCtx.newJsonParser().setPrettyPrint(true).encodeResourceToString(inParams));
+
+		Parameters respParams = ourClient
+			.operation()
+			.onType(ConceptMap.class)
+			.named("translate")
+			.withParameters(inParams)
+			.execute();
+
+		ourLog.info("Response Parameters\n" + myFhirCtx.newJsonParser().setPrettyPrint(true).encodeResourceToString(respParams));
+
+		// Should return v2 since v2 is the most recently updated version.
+		ParametersParameterComponent param = getParameterByName(respParams, "result");
+		assertTrue(((BooleanType) param.getValue()).booleanValue());
+
+		param = getParameterByName(respParams, "message");
+		assertEquals("Matches found!", ((StringType) param.getValue()).getValueAsString());
+
+		assertEquals(1, getNumberOfParametersByName(respParams, "match"));
+		param = getParametersByName(respParams, "match").get(0);
+		assertEquals(3, param.getPart().size());
+		ParametersParameterComponent part = getPartByName(param, "equivalence");
+		assertEquals("equal", ((CodeType) part.getValue()).getValueAsString());
+		part = getPartByName(param, "concept");
+		Coding coding = (Coding) part.getValue();
+		assertEquals("13333", coding.getCode());
+		assertEquals("Target Code 13333", coding.getDisplay());
+		assertFalse(coding.getUserSelected());
+		assertEquals(CS_URL_2, coding.getSystem());
+		assertEquals("Version 2", coding.getVersion());
+		part = getPartByName(param, "source");
+		assertEquals(url, ((UriType) part.getValue()).getValueAsString());
+	}
+
+	@Test
+	public void testTranslateWithVersionedConcaptMapUrl_NoVersion_null_v2() {
+
+		String url = "http://url";
+		createConceptMap(url, "v1", "12222", "Target Code 12222"); 
+		createConceptMap(url, null, "13333", "Target Code 13333"); // second version is null
+		
+		// Call translate with no ConceptMap version.
+		Parameters inParams = new Parameters();
+		inParams.addParameter().setName("url").setValue(new UriType(url));
+		inParams.addParameter().setName("system").setValue(new UriType(CS_URL));
+		inParams.addParameter().setName("targetsystem").setValue(new UriType(CS_URL_2));
+		inParams.addParameter().setName("code").setValue(new CodeType("11111"));
+
+		ourLog.info("Request Parameters:\n" + myFhirCtx.newJsonParser().setPrettyPrint(true).encodeResourceToString(inParams));
+
+		Parameters respParams = ourClient
+			.operation()
+			.onType(ConceptMap.class)
+			.named("translate")
+			.withParameters(inParams)
+			.execute();
+
+		ourLog.info("Response Parameters\n" + myFhirCtx.newJsonParser().setPrettyPrint(true).encodeResourceToString(respParams));
+
+		// Should return v2 since v2 is the most recently updated version.
+		ParametersParameterComponent param = getParameterByName(respParams, "result");
+		assertTrue(((BooleanType) param.getValue()).booleanValue());
+
+		param = getParameterByName(respParams, "message");
+		assertEquals("Matches found!", ((StringType) param.getValue()).getValueAsString());
+
+		assertEquals(1, getNumberOfParametersByName(respParams, "match"));
+		param = getParametersByName(respParams, "match").get(0);
+		assertEquals(3, param.getPart().size());
+		ParametersParameterComponent part = getPartByName(param, "equivalence");
+		assertEquals("equal", ((CodeType) part.getValue()).getValueAsString());
+		part = getPartByName(param, "concept");
+		Coding coding = (Coding) part.getValue();
+		assertEquals("13333", coding.getCode());
+		assertEquals("Target Code 13333", coding.getDisplay());
+		assertFalse(coding.getUserSelected());
+		assertEquals(CS_URL_2, coding.getSystem());
+		assertEquals("Version 2", coding.getVersion());
+		part = getPartByName(param, "source");
+		assertEquals(url, ((UriType) part.getValue()).getValueAsString());
+	}
+	
+	@Test
+	public void testTranslateWithConceptMap_WrongUrl_NoVersion() {
+		
+		String url = "http://url";
+		createConceptMap(url, "v1", "12222", "Target Code 12222");
+		createConceptMap(url, "v2", "13333", "Target Code 13333");
+		
+		Parameters inParams = new Parameters();
+		inParams.addParameter().setName("url").setValue(new UriType("http://invalid.url.com")); // no exsits url
+		inParams.addParameter().setName("system").setValue(new UriType(CS_URL));
+		inParams.addParameter().setName("targetsystem").setValue(new UriType(CS_URL_2));
+		inParams.addParameter().setName("code").setValue(new CodeType("11111"));
+
+		ourLog.info("Request Parameters:\n" + myFhirCtx.newJsonParser().setPrettyPrint(true).encodeResourceToString(inParams));
+
+		Parameters respParams = ourClient
+			.operation()
+			.onType(ConceptMap.class)
+			.named("translate")
+			.withParameters(inParams)
+			.execute();
+
+		ourLog.info("Response Parameters\n" + myFhirCtx.newJsonParser().setPrettyPrint(true).encodeResourceToString(respParams));
+		
+		ParametersParameterComponent param = getParameterByName(respParams, "result");
+		assertFalse(((BooleanType) param.getValue()).booleanValue());
+
+		param = getParameterByName(respParams, "message");
+		assertEquals("No matches found!", ((StringType) param.getValue()).getValueAsString());
+
+	}
+	
+	@Test
+	public void testTranslateWithReverseConceptMapUrlAndVersion() {
 		
 		String url = "http://url";
 		createReverseConceptMap(url, "v1", "12222", "Source Code 12222");
@@ -255,7 +497,7 @@ public class ResourceProviderDstu3ConceptMapTest extends BaseResourceProviderDst
 	}
 	
 	@Test
-	public void testTranslateWithReverseConcaptMapUrl_NoVersion() {
+	public void testTranslateWithReverseConceptMapUrl_NoVersion() {
 		
 		String url = "http://url";
 		createReverseConceptMap(url, "v1", "12222", "Source Code 12222");
@@ -300,7 +542,7 @@ public class ResourceProviderDstu3ConceptMapTest extends BaseResourceProviderDst
 	}
 	
 	@Test
-	public void testTranslateWithReverseConcaptMapUrl_NoVersion_null_v1() {
+	public void testTranslateWithReverseConceptMapUrl_NoVersion_null_v1() {
 		
 		String url = "http://url";
 		createReverseConceptMap(url, null, "12222", "Source Code 12222");
@@ -345,7 +587,7 @@ public class ResourceProviderDstu3ConceptMapTest extends BaseResourceProviderDst
 	}
 	
 	@Test
-	public void testTranslateWithReverseConcaptMapUrl_NoVersion_null_v2() {
+	public void testTranslateWithReverseConceptMapUrl_NoVersion_null_v2() {
 		
 		String url = "http://url";
 		createReverseConceptMap(url, "v1", "12222", "Source Code 12222");
@@ -387,6 +629,26 @@ public class ResourceProviderDstu3ConceptMapTest extends BaseResourceProviderDst
 		assertEquals("Version 1", coding.getVersion());
 		part = getPartByName(param, "source");
 		assertEquals(url, ((UriType) part.getValue()).getValueAsString());	
+	}
+	
+	private void createConceptMap(String url, String version, String targetCode, String targetDisplay) {
+		
+		ConceptMap conceptMap = new ConceptMap();
+		conceptMap.setUrl(url).setVersion(version).setSource(new UriType(VS_URL)).setTarget(new UriType(VS_URL_2));
+
+		ConceptMapGroupComponent group1 = conceptMap.addGroup();
+		group1.setSource(CS_URL).setSourceVersion("Version 1").setTarget(CS_URL_2).setTargetVersion("Version 2");
+
+		SourceElementComponent element1 = group1.addElement();
+		element1.setCode("11111").setDisplay("Source Code 11111");
+
+		TargetElementComponent target1 = element1.addTarget();
+		target1.setCode(targetCode).setDisplay(targetDisplay).setEquivalence(ConceptMapEquivalence.EQUAL);
+
+		IIdType conceptMapId = myConceptMapDao.create(conceptMap, mySrd).getId().toUnqualifiedVersionless();
+		conceptMap = myConceptMapDao.read(conceptMapId);
+		
+		ourLog.info("ConceptMap: \n" + myFhirCtx.newJsonParser().setPrettyPrint(true).encodeResourceToString(conceptMap));
 	}
 	
 	private void createReverseConceptMap(String url, String version, String sourceCode, String sourceDisplay) {
